Rewrote partials, all blocks support standalone
Blocks supporting standalone end tags was a matter of saying that this
could have been the beginning of the content.

Rewrote all partials to do in-bash string processing in order to add the
indent characters to all lines that have content.

Failing 28 out of 123 tests.  At this point I have looked at all of the
tests and feel confident that they are because of features that are
intentionally not supported.