Must have IFS set properly
This lets the value returned from mustache-is-standalone get parsed
correctly.  When this is a standalone template, the value returned looks
like "12 34" where each number is a position.  Later, this is used in a
line that looks like `( $VALUE_RETURNED )`.  When IFS does not contain a
space, this array is not built correctly.